+++ packages/typedoc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/typedoc
 
+## 0.14.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@0.13.4
+
 ## 0.14.3
 
 ### Patch Changes